refactor(allocator): extract routes/public.py blueprint
Moves the participant-facing landing page, seat-assignment endpoint,
and free-seat counter out of main.py into routes/public.py. The four
failure paths in submit_vm_details (idempotent rejoin, no-seats 503,
RotationFailed unhealthy+release+503, and the generic outer catch) are
preserved verbatim with their explanatory comments.
